CPR-1201 added handler for address arrival events from sas
